requirements.txt
openai>=1.60.1
torch>=1.0.0
pillow>=8.0.0
opencv-python>=4.0.0
pyautogui>=0.9.53
pygetwindow>=0.0.9
pytesseract>=0.3.10
mss>=10.0.0
requests>=2.32.3